+++ deepparse/network/bpemb_seq2seq.py
@@ -59,7 +59,7 @@ class BPEmbSeq2SeqModel(Seq2SeqModel):
         if path_to_retrained_model is not None:
             self._load_weights(path_to_retrained_model)
         elif pre_trained_weights:
-            # Means we use the pre-trained weights
+            # Means we use the pretrained weights
             model_weights_name = "bpemb"
             if attention_mechanism:
                 model_weights_name += "_attention"